Optimize structure based on review
