+++ src/redux/history/reducers.ts
@@ -14,8 +14,9 @@ import {
   UpdateEntryAction
 } from './types'
 
-// The history reducer will be created without entries because of async entry retrieval.
-// Entries will be added after reducer initialization.
+// Q: Why is the reducer initialized with an empty array instead of the actual history entries like in the config reducer?
+// A: The history reducer will be created without entries because of async entry retrieval.
+//    Entries will be added after reducer initialization.
 
 export const HistoryReducer: Reducer<HistoryEntry[], HistoryAction> = (state: HistoryEntry[] = [], action: HistoryAction) => {
   switch (action.type) {
